fix(extension): single file background handler regist
