--- apps/extension/lib/single-file/core/content/content-hooks-frames-extension-injection.js
@@ -1,34 +1,0 @@
-/*
- * Copyright 2010-2020 Gildas Lormeau
- * contact : gildas.lormeau <at> gmail.com
- * 
- * This file is part of SingleFile.
- *
- *   The code in this file is free software: you can redistribute it and/or 
- *   modify it under the terms of the GNU Affero General Public License 
- *   (GNU AGPL) as published by the Free Software Foundation, either version 3
- *   of the License, or (at your option) any later version.
- * 
- *   The code in this file is distributed in the hope that it will be useful, 
- *   but WITHOUT ANY WARRANTY; without even the implied warranty of 
- *   MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU Affero 
- *   General Public License for more details.
- *
- *   As additional permission under GNU AGPL version 3 section 7, you may 
- *   distribute UNMODIFIED VERSIONS OF THIS file without the copy of the GNU 
- *   AGPL normally required by section 4, provided you include this license 
- *   notice and a URL through which recipients can access the Corresponding 
- *   Source.
- */
-
-const browser = globalThis.browser;
-const document = globalThis.document;
-const Document = globalThis.Document;
-
-if (document instanceof Document && browser && browser.runtime && browser.runtime.getURL) {
-	const scriptElement = document.createElement("script");
-	scriptElement.src = browser.runtime.getURL("/lib/single-file-hooks-frames.js");
-	scriptElement.async = false;
-	(document.documentElement || document).appendChild(scriptElement);
-	scriptElement.remove();
-}
